% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclass.R
\name{reclass}
\alias{reclass}
\title{Reclass Function}
\usage{
reclass(dat, nlev)
}
\arguments{
\item{dat}{Base de datos en formato "data.frame"}

\item{nlev}{Maximo de niveles para considerar una variable factor}
}
\description{
Convierte variables a clase correspondiente (para data.frame)
}
\examples{
reclass(data(mtcars))
}
\keyword{reclass}

